third_party/blink/web_tests/fast/dom/shadow/all-in-shadow-tree-expected.txt

Test for issue 432257: Matched properties cache should work for all property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(test1.shadowRoot.firstChild).width is "100px"
PASS window.getComputedStyle(test2.shadowRoot.firstChild).width is "100px"
PASS successfullyParsed is true

TEST COMPLETE
